Move LocaleCacheAction command to new admin subpackage
