+++ clock/bot/manager.py
@@ -16,7 +16,7 @@ from bot.action.standard.perchat import PerChatAction
 from bot.bot import Bot
 
 from clock import project_info
-from clock.bot.commands.locale_cache import LocaleCacheAction
+from clock.bot.commands.admin.locale_cache import LocaleCacheAction
 from clock.bot.inline.chosen_result import ChosenInlineResultClockAction
 from clock.bot.inline.query.action import InlineQueryClockAction
